+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - less_than
 - add type guard to atom parser
 - add type guard to binary parser
+- add type guard to boolean parser
 - add type guard to integer parser
 
 # v0.1.0 (Sat Aug 25 21:55:43 2018 -0700)